+++ src/export_vn/download_vn.py
@@ -11,8 +11,16 @@ Properties
 
 """
 import logging
+from collections import deque
 from datetime import datetime, timedelta
+from itertools import chain
+from sys import getsizeof
+from time import perf_counter_ns
 
+try:
+    from reprlib import repr
+except ImportError:
+    pass
 from biolovision.api import (
     EntitiesAPI,
     FamiliesAPI,
@@ -32,6 +40,45 @@ from export_vn.store_postgresql import ReadPostgresql
 from . import _, __version__
 
 logger = logging.getLogger("transfer_vn.download_vn")
+
+
+def total_size(o, handlers={}):
+    """Returns the approximate memory footprint an object and all of its contents.
+
+    Automatically finds the contents of the following builtin containers and
+    their subclasses:  tuple, list, deque, dict, set and frozenset.
+    To search other containers, add handlers to iterate over their contents:
+
+        handlers = {SomeContainerClass: iter,
+                    OtherContainerClass: OtherContainerClass.get_elements}
+
+    """
+    dict_handler = lambda d: chain.from_iterable(d.items())
+    all_handlers = {
+        tuple: iter,
+        list: iter,
+        deque: iter,
+        dict: dict_handler,
+        set: iter,
+        frozenset: iter,
+    }
+    all_handlers.update(handlers)  # user handlers take precedence
+    seen = set()  # track which object id's have already been seen
+    default_size = getsizeof(0)  # estimate sizeof object without __sizeof__
+
+    def sizeof(o):
+        if id(o) in seen:  # do not double count the same object
+            return 0
+        seen.add(id(o))
+        s = getsizeof(o, default_size)
+
+        for typ, handler in all_handlers.items():
+            if isinstance(o, typ):
+                s += sum(map(sizeof, handler(o)))
+                break
+        return s
+
+    return sizeof(o)
 
 
 class DownloadVnException(Exception):
@@ -113,7 +160,9 @@ class DownloadVn:
             i += 1
             log_msg = _("Iteration {}, opt_params = {}").format(i, opt_params)
             logger.debug(log_msg)
+            timing = perf_counter_ns()
             items_dict = self._api_instance.api_list(opt_params=opt_params)
+            timing = (perf_counter_ns() - timing) / 1000
             # Call backend to store generic log
             self._backend.log(
                 self._config.site,
@@ -121,6 +170,8 @@ class DownloadVn:
                 self._api_instance.transfer_errors,
                 self._api_instance.http_status,
                 log_msg,
+                total_size(items_dict),
+                timing,
             )
             # Call backend to store results
             self._backend.store(self._api_instance.controler, str(i), items_dict)
@@ -201,7 +252,9 @@ class Fields(DownloadVn):
             i += 1
             log_msg = _("Iteration {}, opt_params = {}").format(i, opt_params)
             logger.debug(log_msg)
+            timing = perf_counter_ns()
             items_dict = self._api_instance.api_list(opt_params=opt_params)
+            timing = (perf_counter_ns() - timing) / 1000
             # Call backend to store generic log
             self._backend.log(
                 self._config.site,
@@ -209,6 +262,8 @@ class Fields(DownloadVn):
                 self._api_instance.transfer_errors,
                 self._api_instance.http_status,
                 log_msg,
+                total_size(items_dict),
+                timing,
             )
             # Call backend to store groups of fields
             self._backend.store("field_groups", str(i), items_dict)
@@ -262,7 +317,7 @@ class LocalAdminUnits(DownloadVn):
         territorial_unit_ids : list
             List of territorial_units to include in storage.
         """
-        if territorial_unit_ids is not None:
+        if territorial_unit_ids is not None and len(territorial_unit_ids) > 0:
             for id_canton in territorial_unit_ids:
                 logger.debug(
                     _("Getting local_admin_units from id_canton %s, using API list"),
@@ -349,11 +404,13 @@ class Observations(DownloadVn):
                                 id_taxo_group,
                                 specie["id"],
                             )
+                            timing = perf_counter_ns()
                             items_dict = self._api_instance.api_list(
                                 id_taxo_group,
                                 id_species=specie["id"],
                                 short_version=short_version,
                             )
+                            timing = (perf_counter_ns() - timing) / 1000
                             # Call backend to store list by taxo_group, species log
                             self._backend.log(
                                 self._config.site,
@@ -365,6 +422,8 @@ class Observations(DownloadVn):
                                     id_taxo_group,
                                     specie["id"],
                                 ),
+                                total_size(items_dict),
+                                timing,
                             )
                             # Call backend to store results
                             self._backend.store(
@@ -373,9 +432,11 @@ class Observations(DownloadVn):
                                 items_dict,
                             )
                 else:
+                    timing = perf_counter_ns()
                     items_dict = self._api_instance.api_list(
                         id_taxo_group, short_version=short_version
                     )
+                    timing = (perf_counter_ns() - timing) / 1000
                     # Call backend to store list by taxo_group log
                     self._backend.log(
                         self._config.site,
@@ -383,6 +444,8 @@ class Observations(DownloadVn):
                         self._api_instance.transfer_errors,
                         self._api_instance.http_status,
                         (_("observations from taxo_group %s"), id_taxo_group),
+                        total_size(items_dict),
+                        timing,
                     )
                     # Call backend to store results
                     self._backend.store(
@@ -470,10 +533,14 @@ class Observations(DownloadVn):
                             q_param["entry_date"] = "1"
                         else:
                             q_param["entry_date"] = "0"
-                    if territorial_unit_ids is None:
+                    if territorial_unit_ids is None or len(territorial_unit_ids) == 0:
                         t_us = self._t_units
                     else:
-                        t_us = [u for u in self._t_units if u[0]["short_name"] in territorial_unit_ids]
+                        t_us = [
+                            u
+                            for u in self._t_units
+                            if u[0]["short_name"] in territorial_unit_ids
+                        ]
                     for t_u in t_us:
                         logger.debug(
                             _(
@@ -486,9 +553,12 @@ class Observations(DownloadVn):
                             t_u[0]["id_country"] + t_u[0]["short_name"]
                         ]
 
+                        timing = perf_counter_ns()
                         items_dict = self._api_instance.api_search(
                             q_param, short_version=short_version
                         )
+                        timing = (perf_counter_ns() - timing) / 1000
+
                         # Call backend to store results
                         nb_o = self._backend.store(
                             self._api_instance.controler,
@@ -515,6 +585,8 @@ class Observations(DownloadVn):
                             self._api_instance.transfer_errors,
                             self._api_instance.http_status,
                             log_msg,
+                            total_size(items_dict),
+                            timing,
                         )
                         logger.info(log_msg)
                     seq += 1
@@ -681,43 +753,50 @@ class Observations(DownloadVn):
                 logger.error(
                     _("No date found for last download, increment not performed")
                 )
-            print(updated)
+            # print(updated)
 
             # Process updates
             if len(updated) > 0:
                 log_msg = _("Creating or updating {} observations").format(len(updated))
                 logger.debug(log_msg)
                 # Update backend store, in chunks
-                [
+                for i in range(
+                    (len(updated) + self._config.tuning_max_list_length - 1)
+                    // self._config.tuning_max_list_length
+                ):
+                    s_list = ",".join(
+                            updated[
+                                i
+                                * self._config.tuning_max_list_length : (i + 1)
+                                * self._config.tuning_max_list_length
+                            ]
+                        )
+                    logger.debug(_("Updating slice %s"), s_list)
+                    timing = perf_counter_ns()
+                    items_dict = self._api_instance.api_list(
+                        taxo,
+                        id_sightings_list=s_list,
+                        short_version=short_version,
+                    )
+                    timing = (perf_counter_ns() - timing) / 1000
+
                     # Call backend to store results
                     self._backend.store(
                         self._api_instance.controler,
-                        str(id_taxo_group) + "_1",
-                        self._api_instance.api_list(
-                            taxo,
-                            id_sightings_list=",".join(
-                                updated[
-                                    i
-                                    * self._config.tuning_max_list_length : (i + 1)
-                                    * self._config.tuning_max_list_length
-                                ]
-                            ),
-                            short_version=short_version,
-                        ),
+                        str(id_taxo_group) + "_upd_" + str(i),
+                        items_dict,
                     )
-                    for i in range(
-                        (len(updated) + self._config.tuning_max_list_length - 1)
-                        // self._config.tuning_max_list_length
+
+                    # Call backend to store log
+                    self._backend.log(
+                        self._config.site,
+                        self._api_instance.controler,
+                        self._api_instance.transfer_errors,
+                        self._api_instance.http_status,
+                        log_msg,
+                        total_size(items_dict),
+                        timing,
                     )
-                ]
-                # Call backend to store log
-                self._backend.log(
-                    self._config.site,
-                    self._api_instance.controler,
-                    self._api_instance.transfer_errors,
-                    self._api_instance.http_status,
-                    log_msg,
-                )
 
             # Process deletes
             if len(deleted) > 0: